Import modules in `teq` as late as possible.
If we're running the Web server, there's no need to import db from the
top-level script: the web app itself will manage database connections.
Conversely, if we're not running the Web server, there's no need to import
the webapp module. Finally, if we don't need the Core module, don't import
it at all: it causes problems on current versions of FreeBSD.